Green

+++ sortega/java/src/test/java/katapotter/PriceTest.java
@@ -2,12 +2,17 @@ package katapotter;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
+
+import static org.junit.Assert.assertEquals;
 
 public class PriceTest {
 
     @Test()
     public void noBooksNoMoney() {
-        assertEquals(0, Price.price(Arrays.asList()));
+        assertEquals(BigDecimal.valueOf(0), Price.price(Arrays.asList()));
     }
+
+
 }
